Fix invalid signature on authenticated GET request
The signature for GET requests must be made with the path+arg string
and not just the path. Prior to this patch authenticated GET requests
would fail with an "invalid signature" error:

iex(19)> ExGdax.list_fills(%{product_id: "BTC-EUR"})
{:error, "invalid signature", 400}

This patch ensures that the signature is created with the query string,
not just the URI path. `list_fills`, as above, then returns data.